project/requirements.txt
asgiref==3.8.1
Django==5.0.6
django-jazzmin==3.0.0
django-phonenumber-field==7.3.0
djangorestframework==3.15.1
phonenumbers==8.13.38
sqlparse==0.5.0
tzdata==2024.1
